+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - The Cache Folder Path column now displays aliases in the cached file table in the File Storage utility.
 - Improved error handling in a previous migration.
+- Made some minor optimisations to the file storage driver.
 
 ## 4.10.3 - 2024-01-13